+++ notionrs/tests/mutable/data_source/mod.rs
@@ -1,6 +1,3 @@
 mod crud_data_source;
 mod list_data_source_templates;
-mod query_data_source;
-mod query_data_source_all;
-mod query_data_source_with_struct;
 mod retrieve_data_source;
