+++ notionrs/tests/readonly/data_source/query_data_source.rs
@@ -2,33 +2,26 @@ mod integration_tests {
 
     // # --------------------------------------------------------------------------------
     //
-    // query_data_source_all
+    // query_data_source
     //
     // # --------------------------------------------------------------------------------
 
-    use futures::TryStreamExt;
-    use notionrs::r#trait::PaginateExt;
     use notionrs_types::prelude::*;
 
+    static DATA_SOURCE_ID: &str = "33da03d7-9b26-81cb-90c7-000b8fb827a8";
+
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
-        let response: Vec<PageResponse> = client
+        let res = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
-        println!("{}", serde_json::to_string(&response)?);
+            .data_source_id(DATA_SOURCE_ID)
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
+        println!("{}", serde_json::to_string(&res)?);
 
         Ok(())
     }
@@ -40,24 +33,18 @@ mod integration_tests {
     // # --------------------------------------------------------------------------------
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_page_size() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_page_size() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
-        let response: Vec<PageResponse> = client
+        let res = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
-        println!("{}", serde_json::to_string(&response)?);
+            .data_source_id(DATA_SOURCE_ID)
+            .page_size(1)
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
+        println!("{}", serde_json::to_string(&res)?);
 
         Ok(())
     }
@@ -69,15 +56,10 @@ mod integration_tests {
     // # --------------------------------------------------------------------------------
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_simple() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_simple() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::date_before(
@@ -85,14 +67,14 @@ mod integration_tests {
             "2024-07-01",
         );
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -100,15 +82,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_checkbox() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_checkbox() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -116,14 +93,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::checkbox_is_not_checked("Checkbox"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -131,15 +108,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_date_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_date_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -174,14 +146,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::date_this_week("Created time"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -189,15 +161,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_files_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_files_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -205,14 +172,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::files_is_not_empty("Files & media"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -221,7 +188,7 @@ mod integration_tests {
 
     // #[tokio::test]
     // #[serial_test::serial]
-    // async fn query_data_source_all_filter_formula_filter() -> Result<(), notionrs::Error> {
+    // async fn query_data_source_filter_formula_filter() -> Result<(), notionrs::Error> {
     //     dotenvy::dotenv().ok();
     //     dotenvy::from_path(std::path::Path::new("../.env"))
     //         .expect("Failed to load ../.env file");
@@ -242,11 +209,11 @@ mod integration_tests {
     //     ]);
 
     //     let request = client
-    //         .query_data_source_all()
-    //         .data_source_id(data_source_id)
+    //         .query_data_source()
+    //         .data_source_id(DATA_SOURCE_ID)
     //         .filter(filter);
 
-    //     let response = request.send().await?;
+    //     let response = request.send::<std::collections::HashMap<String,PageProperty>>().await?;
 
     //     println!("{}", serde_json::to_string(&response)?);
 
@@ -254,15 +221,10 @@ mod integration_tests {
     // }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_multi_select_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_multi_select_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -280,14 +242,14 @@ mod integration_tests {
             ),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -295,15 +257,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_number_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_number_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -321,14 +278,14 @@ mod integration_tests {
             ),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -336,15 +293,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_people_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_people_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -360,14 +312,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::people_is_not_empty("User"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -375,15 +327,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_phone_number_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_phone_number_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -417,14 +364,14 @@ mod integration_tests {
             ),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -432,15 +379,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_relation_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_relation_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -456,14 +398,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::relation_is_not_empty("Relation"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -471,15 +413,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_rollup_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_rollup_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -497,14 +434,14 @@ mod integration_tests {
             ),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -512,15 +449,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_rich_text_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_rich_text_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -538,14 +470,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::rich_text_starts_with("Text", "0"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -553,15 +485,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_select_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_select_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -573,14 +500,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::select_is_not_empty("Select"),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -589,7 +516,7 @@ mod integration_tests {
 
     // #[tokio::test]
     // #[serial_test::serial]
-    // async fn query_data_source_all_filter_status_filter() -> Result<(), notionrs::Error> {
+    // async fn query_data_source_filter_status_filter() -> Result<(), notionrs::Error> {
     //     dotenvy::dotenv().ok();
     //     dotenvy::from_path(std::path::Path::new("../.env"))
     //         .expect("Failed to load ../.env file");
@@ -606,11 +533,11 @@ mod integration_tests {
     //     ]);
 
     //     let request = client
-    //         .query_data_source_all()
-    //         .data_source_id(data_source_id)
+    //         .query_data_source()
+    //         .data_source_id(DATA_SOURCE_ID)
     //         .filter(filter);
 
-    //     let response = request.send().await?;
+    //     let response = request.send::<std::collections::HashMap<String,PageProperty>>().await?;
 
     //     println!("{}", serde_json::to_string(&response)?);
 
@@ -618,15 +545,10 @@ mod integration_tests {
     // }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_timestamp_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_timestamp_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -646,14 +568,14 @@ mod integration_tests {
             notionrs_types::object::request::filter::Filter::timestamp_this_week(),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -661,15 +583,10 @@ mod integration_tests {
     }
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_filter_unique_id_filter() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_filter_unique_id_filter() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let filter = notionrs_types::object::request::filter::Filter::or(vec![
@@ -685,14 +602,14 @@ mod integration_tests {
             ),
         ]);
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .filter(filter)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .filter(filter);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
 
@@ -706,29 +623,24 @@ mod integration_tests {
     // # --------------------------------------------------------------------------------
 
     #[tokio::test]
-    #[serial_test::serial]
-    async fn query_data_source_all_sort() -> Result<(), notionrs::Error> {
-        dotenvy::dotenv().ok();
-        dotenvy::from_path(std::path::Path::new("../.env")).expect("Failed to load ../.env file");
+    async fn query_data_source_sort() -> Result<(), notionrs::Error> {
+        dotenvy::from_path(std::path::Path::new(".env.readonly")).ok();
 
-        let data_source_id =
-            std::env::var("NOTION_IT_DATA_SOURCE_ID").unwrap_or_else(|_| String::new());
-
-        let notion_api_key = std::env::var("NOTION_TOKEN").unwrap();
+        let notion_api_key = std::env::var("NOTION_API_KEY").unwrap();
         let client = notionrs::Client::new(notion_api_key);
 
         let sorts = vec![notionrs_types::object::request::sort::Sort::asc(
             "Created time",
         )];
 
-        let response: Vec<PageResponse> = client
+        let request = client
             .query_data_source()
-            .data_source_id(data_source_id)
-            .sorts(sorts)
-            .into_stream()
-            .try_collect()
-            .await
-            .unwrap();
+            .data_source_id(DATA_SOURCE_ID)
+            .sorts(sorts);
+
+        let response = request
+            .send::<std::collections::HashMap<String, PageProperty>>()
+            .await?;
 
         println!("{}", serde_json::to_string(&response)?);
