+++ notionrs/tests/readonly/mod.rs
@@ -1,5 +1,6 @@
 mod block;
 mod custom_emoji;
+mod data_source;
 mod file_upload;
 mod page;
 mod search;
